requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
redis
mysqlclient
python-dotenv
alembic
pytest
pymysql
cryptography
httpx
pytest-asyncio